// src/models/enum/Roles.ts
export enum Roles {
  SUPERADMIN = 'SUPERADMIN', //READ WRITE ALL TABEL ALL RECORDS
  ADMIN = 'ADMIN', //READ WRITE ALL TABLE AND RECORDS ONLY ASSOCIATE FRANCHISE
  READ = 'READ', //ONLY READALL TABLE
  EMP = 'EMP', // LIMITED PRIVILEDGE
  INVENTORYREAD = 'INVENTORYREAD', // ONLY ABLE TO READ ITEMS AND ITS SELLING PRICE
  CLIENT = 'CLIENT', // ONLY ABLE TO READ ITEMS AND ITS SELLING PRICE NO OTHER CHANGE
  ZERO = 'ZERO',
}